requirements.txt
asgiref==3.7.2
Django==3.2.23
django-model-utils==4.3.1
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
